+++ javascript/packages/eslint-config-chotot-base/index.js
@@ -2,7 +2,6 @@ module.exports = {
   extends: [
     './rules/best-practices',
     './rules/errors',
-    './rules/node',
     './rules/style',
     './rules/variables',
     './rules/es6',
@@ -11,7 +10,7 @@ module.exports = {
     './rules/chotot',
   ].map(require.resolve),
   parserOptions: {
-    ecmaVersion: 2020,
+    ecmaVersion: 2021,
     sourceType: 'module',
   },
   rules: {},