+++ javascript/packages/eslint-config-chotot/index.js
@@ -5,9 +5,9 @@ module.exports = {
     './rules/react-hooks',
     './rules/react-a11y',
   ].map(require.resolve),
-  parser: 'babel-eslint',
+  parser: '@babel/eslint-parser',
   parserOptions: {
-    ecmaVersion: 2020,
+    ecmaVersion: 2021,
     sourceType: 'module',
   },
   rules: {}