+++ javascript/packages/eslint-config-chotot/test/test-react-order.js
@@ -1,26 +1,29 @@
 import test from 'tape';
-import { CLIEngine } from 'eslint';
+import { ESLint } from 'eslint';
 import eslintrc from '..';
 import reactRules from '../rules/react';
 import reactA11yRules from '../rules/react-a11y';
 
-const cli = new CLIEngine({
+const eslint = new ESLint({
   useEslintrc: false,
   baseConfig: eslintrc,
 
-  rules: {
-    // It is okay to import devDependencies in tests.
-    'import/no-extraneous-dependencies': [2, { devDependencies: true }],
-    // this doesn't matter for tests
-    'lines-between-class-members': 0,
-  },
+  overrideConfig: {
+    rules: {
+      // It is okay to import devDependencies in tests.
+      'import/no-extraneous-dependencies': [2, { devDependencies: true }],
+      // this doesn't matter for tests
+      'lines-between-class-members': 0,
+      // let's skip this in tests
+      'react/no-unused-class-component-methods': 0
+    },
+  }
 });
 
-function lint(text) {
-  // @see https://eslint.org/docs/developer-guide/nodejs-api.html#executeonfiles
-  // @see https://eslint.org/docs/developer-guide/nodejs-api.html#executeontext
-  const linter = cli.executeOnText(text);
-  return linter.results[0];
+async function lint(text) {
+  // @see https://eslint.org/docs/developer-guide/nodejs-api#-eslintlinttextcode-options
+  const results = await eslint.lintText(text);
+  return results[0];
 }
 
 function wrapComponent(body) {
@@ -40,9 +43,9 @@ test('validate react prop order', (t) => {
     t.deepEqual(reactA11yRules.plugins, ['jsx-a11y', 'react']);
   });
 
-  t.test('passes a good component', (t) => {
+  t.test('passes a good component', async (t) => {
     t.plan(3);
-    const result = lint(wrapComponent(`
+    const result = await lint(wrapComponent(`
   componentDidMount() {}
   setFoo() {}
   getFoo() {}
@@ -57,9 +60,9 @@ test('validate react prop order', (t) => {
     t.notOk(result.errorCount, 'no errors');
   });
 
-  t.test('order: when random method is first', (t) => {
+  t.test('order: when random method is first', async (t) => {
     t.plan(2);
-    const result = lint(wrapComponent(`
+    const result = await lint(wrapComponent(`
   someMethod() {}
   componentDidMount() {}
   setFoo() {}
@@ -73,9 +76,9 @@ test('validate react prop order', (t) => {
     t.deepEqual(result.messages.map((msg) => msg.ruleId), ['react/sort-comp'], 'fails due to sort');
   });
 
-  t.test('order: when random method after lifecycle methods', (t) => {
+  t.test('order: when random method after lifecycle methods', async (t) => {
     t.plan(2);
-    const result = lint(wrapComponent(`
+    const result = await lint(wrapComponent(`
   componentDidMount() {}
   someMethod() {}
   setFoo() {}